+++ CHANGELOG.md
@@ -112,3 +112,5 @@
 
 ## v1.1.0
 * Rewrote syntax to be similar to RoRML.
+* Now displays "Modded" under the version number at top-left corner.
+* Added Skill module